% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable1.R
\name{.y(.x)}
\alias{.y(.x)}
\title{helper function call `.y(.x)` for functional programming, replacing `purrr::at_depth(.x, 0, .y, ...)` which is deprecated.}
\usage{
".y(.x)"(.x, .y, ...)
}
\arguments{
\item{.x}{The first argument for function/formula passed as `.y`.}

\item{.y}{The function/formula to be called with argument `.x`.}

\item{...}{Additional arguments to `.y()`.}
}
\value{
Whatever is returned from `.y(.x, ...)`.
}
\description{
helper function call `.y(.x)` for functional programming, replacing `purrr::at_depth(.x, 0, .y, ...)` which is deprecated.
}
