% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_table1.R
\name{remove_dups}
\alias{remove_dups}
\title{\code{remove_dups}  replaces duplicate information with NA for each variable, to be called inside `dplyr::mutate`.}
\usage{
remove_dups(x)
}
\arguments{
\item{x}{A column in a data frame.}
}
\value{
A data frame column with duplicate information replaced by NA.
}
\description{
\code{remove_dups}  replaces duplicate information with NA for each variable, to be called inside `dplyr::mutate`.
}
