% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytable1.R
\name{tidytable1}
\alias{tidytable1}
\title{Creates Table1 Descriptive Statistics with tidyverse}
\usage{
tidytable1(input_df, info_cols = list(), calc_cols = list(`#missing` =
  function(x) sum(is.na(x))), num_cols = list(mean = mean, sd = sd),
  col_order = c(), row_order = list(), digits = 2,
  add_cat_header_row = TRUE)
}
\arguments{
\item{input_df}{A data_frame for which table1 (the descriptive stats) is produced.}

\item{info_cols}{A list that will be used for informational columns in table1, for example, the label (description) and data source of variables. The names of the list will be used as column names in table1, while the names of value vector will be matched to column names in input_df.}

\item{calc_cols}{A list of table1 columns to be calculated from functions.}

\item{num_cols}{A list of table1 columns to be calculated from functions for numeric variables.}

\item{col_order}{A vector according to which table1 columns will be ordered.}

\item{row_order}{A list according to which table1 rows (variables in input_df) will be ordered.}

\item{digits}{Number of digits for stats in table1 (Default 2).}

\item{add_cat_header_row}{Whether to add a header row for categorical variable (Default TRUE).}
}
\value{
A data frame containing the descriptive stats, by default containing mean and sd for numeric variables, Category, frequency, and `%` for categorical variables, and #missing for all variables.
}
\description{
\code{tidytable1} produces a tidy data frame that contains the common information included in Table 1 of academic papers for an input data frame. Inspired by [tableone](https://cran.r-project.org/web/packages/tableone/), tidytable1 uses the tidyverse functions for easy customization.
}
